refactor(parser): move input parsers to dedicated parser package
